requirements.txt
altgraph==0.17
asgiref==3.3.1
auto-py-to-exe==2.8.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-bootstrap-modal-forms==1.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
Eel==0.12.4
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
idna==2.10
oauthlib==3.1.0
Pillow==8.2.0
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.4
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.3.0
